[Update] 체력, 짜증게이지등 스탯 작성, 변신, 변신해제 InputSystem연동

+++ Assets/Script/Data/GameData/PcData.cs
@@ -6,7 +6,7 @@ public class PcData : IDataRepository
     public string ID;
     public string description;
     public string pcType;
-    public float pcHp;
+    public int pcHp;
     public float pcBasePower;
     public float pcMoveSpeed;
     public string moveSet;
